Extract blocks to their own module
